refactor(pwm): split module into channel, pads, register and signal
